// src/app/components/toast-notification/config/enums/toast.enum.ts
export enum ToastStatusValue {
    INACTIVE = 'inactive',
    ACTIVE = 'active',
    REMOVED = 'removed',
}

export enum ToastDisableTimeOut {
    TIME_OUT = 'timeOut',
    EXTENDED_TIME_OUT = 'extendedTimeOut',
}


export enum ToastProgressAnimation {
    INCREASING = 'increasing',
    DECREASING = 'decreasing',
}